dbus: change 'system-release' dependency to 'fedora-release'
As it turns out, 'fedora-release' only has:

    'Provides: system-release'

rather than:

    'Provides: system-release = %{version}-%{release}'

This means, any explicit dependencies on 'system-release' ignore
versioning completely. This is unfortunate, and breaks our reliance on
'system-release'.

Fortunately, unlike recommended on the wiki, it is safe to depend on
'fedora-release' explicitly. Other fedora flavors have this package as
well, so there is no reason not to.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>